HUD player

+++ Solution/MarriageJam/Assets/Scripts/Damage.cs
@@ -2,9 +2,10 @@
 using UnityEngine;
 
 [Serializable]
-public class Damage  {
+public class Damage
+{
     public string Name = "default";
-    public float Value = 1;
+    public int Value = 1;
     public bool Knockout = false;
     public bool AttackFromRight = false;
 }
